pyproject.toml
[tool.poetry]
name = "goit-web-pd-homework-7"
version = "0.1.0"
description = ""
authors = ["pavlodubovyi <133531274+pavlodubovyi@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.29"
aiosqlite = "^0.20.0"
greenlet = "^3.0.3"
faker = "^24.9.0"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
python-dateutil = "^2.9.0.post0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
